Fixing an error in the previous proof of optimal solution existence
I discovered a problem with the large set expansion theorem in the new
section. However, I was able to remove it and fix it for the case when
a=c. I don't know how to get around it in the general case, so that's
probably something we want to talk about.