First Attempt at generalizing to a≠c
I was able to generalize the (a,a) case we had to (a,am) for any
integer m. It's not hard to see that (am,a) is also possible, but it's
worthless for our purposes. Being a slight bit more clever, I think
it's possible to get general (a,c). The issue is that this analysis is
for G_{n,p} not d-out and I don't think I can keep it simple if we want
to do this for d-out. So that's a bit sad. We should talk about what to
do with that.