Better way of writing the section on optimal recommendation graphs
This version moves the recommendation graphs to the main section,
proves the existence of optimal (c,a)-recommendation subgraphs with
$p=O(log(n)/n)$ edge probabilities, and gives (1-\epsilon)
approximation based on this analysis. The (1-\epsilon) approximation
can even be parallelized (because now we're using matchings)! Sri, I
still have reservations about the d-out model because then the edge
probability depends on the size of R in the underlying graph. With this
model, the size of R is irrelevant as long as it's big enough. I prefer
it. But again, I'm down to change it if we need to in order to better
connect it to the rest of the paper.